% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{calculate bias in bootstrapped proportions-at-length or -age}
\usage{
bias(sim_props, og_props)
}
\arguments{
\item{sim_props}{list of replicated abundance at length or age}

\item{og_props}{original abundance at length or age (computed with data that has not been resampled)}
}
\description{
Computes the mean bias in bootstrapped population at length or age compared to original unsampled
population at length or age
}
