% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_length.R
\name{bias_length}
\alias{bias_length}
\title{calculate bias in bootstrapped proportions-at-length}
\usage{
bias_length(sim_length_props, og_length_props)
}
\arguments{
\item{sim_length_props}{list of replicated abundance at length}

\item{og_length_props}{original abundance at length (computed with data that has not been resampled)}
}
\description{
Computes the mean bias in bootstrapped population at length compared to original unsampled
population at length
}
