% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_length.R
\name{bin_length}
\alias{bin_length}
\title{Bin the length data}
\usage{
bin_length(length_DT, bin = NULL)
}
\arguments{
\item{length_DT}{length frequency input data frame.}

\item{bin}{bin size (default = 1 cm), also can use custom length bins through defining vector of upper length bin limits, i.e., c(5, 10, 20, 35, 60), plus length group will automatically be populated and denoted with the largest defined bin + 1 (i.e., 61 for provided example)}
}
\value{
Data frame of length adapted to length bin vector
}
\description{
Bin the length data in user-defined vector
}
