% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_age.R
\name{boot_age}
\alias{boot_age}
\title{Resample age frequency data w/replacement.}
\usage{
boot_age(age_DT)
}
\arguments{
\item{age_DT}{age frequency input data frame. Can be the output object from boot_trip() if boot.trip == true, or boot_haul() if boot.haul == true, or both}
}
\value{
age frequency data table including resampled age frequencies as AGE and associated data
}
\description{
Resample age frequency data w/replacement.
}
