% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_haul.R
\name{boot_haul}
\alias{boot_haul}
\title{Resample hauls w/replacement.}
\usage{
boot_haul(length_DT)
}
\arguments{
\item{length_DT}{length frequency input data table. Can be the output object from boot_trip() if boot.trip == true}
}
\value{
length frequency data table including resampled hauls as HAUL_JOIN and associated data
}
\description{
Resample hauls w/replacement.
}
