% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_length.R
\name{boot_length}
\alias{boot_length}
\title{Resample length frequency data w/replacement.}
\usage{
boot_length(resampled_hauls)
}
\arguments{
\item{resampled_hauls}{resampled hauls input data table. Data table format input same as length_DT but with only resampled hauls and associated data, plus a unique identifier for resampled HAUL_JOIN called hauljoin_unique}
}
\value{
data table of resampled length frequencies by YEAR, HAUL_JOIN. Data table format output same as length_DT to be ready for expansion.
}
\description{
Resample length frequency data w/replacement.
}
