% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_trip.R
\name{boot_trip}
\alias{boot_trip}
\title{Resample trips w/replacement.}
\usage{
boot_trip(length_DT)
}
\arguments{
\item{length_DT}{length frequency input data table.}
}
\value{
length frequency data table including resampled trips as TRIP_JOIN and associated data
}
\description{
Resample trips w/replacement.
}
