% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_post_strata.R
\name{combine_post_strata}
\alias{combine_post_strata}
\title{Combine post-stratified ISS results. Under Development.}
\usage{
combine_post_strata(freq_data, out_stats)
}
\arguments{
\item{freq_data}{length or age frequency input data frame}

\item{out_stats}{output object from fishery_iss()}
}
\value{
out_stats object to be returned by fishery_iss() that includes one vector of ISS based on the weighted average between post-strata
}
\description{
Combine the post-stratified ISS results by post-strata fish weight or numbers as weighted average. Includes evidence somewhere (DECIDE BEST PLACE) that states it was combined by post_strata and reports what the strata were and the method used to combine them.
}
