% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_sample.R
\name{comp_sample}
\alias{comp_sample}
\title{Resample with replacement a set of length composition observations of the same length}
\usage{
comp_sample(length_comps, n.samples)
}
\arguments{
\item{length_comps}{vector of length measurements}

\item{n.samples}{number of samples to resample}
}
\value{
vector of resampled lengths of length n.samples
}
\description{
Modified base::sample() function to accept n=1 without undesirable treatment:  base::sample() outputs sample(1:n) when n=1 for some reason.
}
