% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_akfin_data.R
\name{download_akfin_data}
\alias{download_akfin_data}
\title{Download data from AKFIN needed to run the primary function in baseISS called fishery_iss()}
\usage{
download_akfin_data(
  akfin_credentials,
  species_code,
  composition_type = NULL,
  save_file_path = NULL
)
}
\arguments{
\item{akfin_credentials}{list(uid = userID, pwd = password)}

\item{species_code}{numeric}

\item{composition_type}{"LENGTH", or "AGE" in caps, or NULL. NULL retrieves haul data only and will require joining to your length or age data to run fishery_iss.R.}

\item{save_file_path}{character path including file name, preferably a .RDS file. Save it somewhere confidential. If NULL, doesn't save anything.}
}
\value{
tidytable data frame ready for use in baseISS
}
\description{
Connect to AKFIN data base and download species-specific composition data or just haul data to join to composition data required to run fishery_ISS.R.
}
