% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_by_sampling_strata.R
\name{expand_by_sampling_strata}
\alias{expand_by_sampling_strata}
\title{Expand by observer sampling strata}
\usage{
expand_by_sampling_strata(freq_data)
}
\arguments{
\item{freq_data}{length or age frequency input data frame}
}
\value{
length frequency data table with modified WEIGHT1 column incorporating sampling strata weighting
}
\description{
expands the length frequency data considering observer sampling strata. The length samples are weighted proportional to the number of fish caught in each stratum.
}
