% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_length_props.R
\name{expand_length_props}
\alias{expand_length_props}
\title{Expand sample level proportions-at-length to population level}
\usage{
expand_length_props(length_DT)
}
\arguments{
\item{length_DT}{data table of length frequency data in the form of EBS Pcod y2 object}
}
\description{
takes length frequency data and expands it into fished population proportions at length using the weightings from the relevant data. The relevant data can be the og data or bootstrapped data.
}
