% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_props.R
\name{expand_props}
\alias{expand_props}
\title{Expand species-specific sample level proportions-at-length or -age to population level}
\usage{
expand_props(
  freq_data,
  species_code,
  area_code,
  length_based = TRUE,
  boot.length = FALSE,
  expand.by.sampling.strata = FALSE,
  expansion_factors = expansion_factors
)
}
\arguments{
\item{freq_data}{length or age frequency input data frame}

\item{species_code}{species specific number code to apply the expansion for. Conditional statements below activate species-specific expansions.}

\item{area_code}{area specific character code to apply the conditional expansion below.}

\item{length_based}{Boolean. If TRUE, then calculate length iss. if FALSE, then calculate age iss.}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length. In this function, this activates a switch to calculate SUM_FREQUENCY for og_props.}

\item{expand.by.sampling.strata}{MAKE THIS GENERIC IN FUTURE. expand by observer sampling strata? If TRUE, then an additional weighting factor is calculated and applied to WEIGHT1 based on the number of fish caught in each sampling stratum.}

\item{expansion_factors}{expansion weighting factors to apply to the proportions. If NULL, then no expansion factors are applied. Otherwise, the conditional options coded in expand_props.R are "haul_numbers" or "haul_numbers" and "month_numbers". Consider improving/generalizing this by calling it expansion_weighting_factors = list(type = c("weight", "number"), factors = c("haul", "area", "month", "gear", etc.) to give the user the option of what aspects (columns) of the data to expand by and do it by weight of fish or number of fish in those categories.}
}
\description{
This is intended to be a large function that consists of many conditional statements, each of which pertains to a unique species/area expansion method. The internals of the conditional statements can eventually turn into many species/area/length/age specific functions housed in another script. Takes length or age frequency data and expands it into fished population proportions at length or age using the weightings from the relevant data. The relevant data can be the og data or bootstrapped data.
}
