% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_iss.R
\name{fishery_iss}
\alias{fishery_iss}
\title{Fishery input sample size function}
\usage{
fishery_iss(iters = 1, lfreq_data, yrs = NULL, boot_thl = FALSE)
}
\arguments{
\item{iters}{number of iterations}

\item{lfreq_data}{length frequency input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{boot_thl}{Boolean. Resample trips, hauls, and lengths w/replacement? (default = FALSE). FALSE will return og proportions-at-length}
}
\value{
Dataframe of input sample size by year
}
\description{
Bootstrap data sources to replicate EBS Pcod fishery observer measurements of length composition
for computation of input sample size.
Follows loosely srvy_iss.R from surveyISS package.
}
