% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_length_props.R
\name{fishery_length_props}
\alias{fishery_length_props}
\title{Resample data and compute proportions-at-length population expansion}
\usage{
fishery_length_props(
  lfreq_data,
  yrs = NULL,
  boot.trip = FALSE,
  boot.haul = FALSE,
  boot.length = FALSE,
  expand.by.sampling.strata = FALSE,
  expand_using_weighting_factors = expand_using_weighting_factors
)
}
\arguments{
\item{lfreq_data}{length frequency input dataframe, not necessarily a data.table or tidytable object yet, hence a different name than subsequent functions (lengthDT)}

\item{yrs}{any year filter >= (default = NULL)}

\item{boot.trip}{Boolean. Resample trips w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.haul}{Boolean. Resample hauls w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{expand.by.sampling.strata}{expand by observer sampling strata? If TRUE, then an additional weighting factor is calculated and applied to WEIGHT1 based on the number of fish caught in each sampling stratum.}

\item{expand_using_weighting_factors}{expand using weighting factors? If TRUE, then then "WEIGHT2" and "WEIGHT4" are applied.}
}
\value{
List of a dataframe of annual population proportions-at-length .lpop.
}
\description{
Follows loosely srvy_comps.R from surveyISS package.
}
