% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_props.R
\name{fishery_props}
\alias{fishery_props}
\title{Resample data and compute proportions-at-length or -age population expansion}
\usage{
fishery_props(
  species_code,
  area_code,
  length_based = TRUE,
  freq_data,
  yrs = NULL,
  boot.trip = FALSE,
  boot.haul = FALSE,
  boot.length = FALSE,
  boot.age = FALSE,
  expand.by.sampling.strata = FALSE,
  expansion_factors = expansion_factors
)
}
\arguments{
\item{species_code}{species number code. Used for specific expansion. This is not (yet) an input data filter, it is only for output naming convention and to condition on expansion method.}

\item{area_code}{area character code. Used for specific expansion. This is not (yet) an input data filter, it is only for output naming convention and to condition on expansion method.}

\item{length_based}{Boolean. If TRUE, then calculate length iss. if FALSE, then calculate age iss.}

\item{freq_data}{length or age frequency input data frame}

\item{yrs}{any year filter >= (default = NULL)}

\item{boot.trip}{Boolean. Resample trips w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length or -age}

\item{boot.haul}{Boolean. Resample hauls w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length or -age}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.age}{Boolean. Resample ages w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-age}

\item{expand.by.sampling.strata}{expand by observer sampling strata? If TRUE, then an additional weighting factor is calculated and applied to WEIGHT1 based on the number of fish caught in each sampling stratum.}

\item{expansion_factors}{expansion weighting factors to apply to the proportions. If NULL, then no expansion factors are applied. Otherwise, the conditional options coded in expand_props.R are "haul_numbers" or "haul_numbers" and "month_numbers". Consider improving/generalizing this by calling it expansion_weighting_factors = list(type = c("weight", "number"), factors = c("haul", "area", "month", "gear", etc.) to give the user the option of what aspects (columns) of the data to expand by and do it by weight of fish or number of fish in those categories.}
}
\value{
List of a dataframe of annual population proportions-at-length .lpop.
}
\description{
Follows loosely srvy_comps.R from surveyISS package.
}
