% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_sample.R
\name{generic_sample}
\alias{generic_sample}
\title{Resample with replacement a vector of the same length. More generic than simply using base::sample}
\usage{
generic_sample(flex.vec, n.samples)
}
\arguments{
\item{flex.vec}{flexible (can be length 1) vector of input elements}

\item{n.samples}{number of elements in the return sample}
}
\value{
vector of resampled elements of length n.samples
}
\description{
Modified base::sample() function to accept n=1 without undesirable treatment:  base::sample() outputs sample(1:n) when n=1 for some reason.
}
