% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_sample_change_length_N.R
\name{generic_sample_change_length_N}
\alias{generic_sample_change_length_N}
\title{Resample with replacement a vector of user-defined length. More generic than simply using base::sample}
\usage{
generic_sample_change_length_N(flex.vec, new_length_N = NULL)
}
\arguments{
\item{flex.vec}{flexible (can be length 1) vector of input elements}

\item{new_length_N}{user-defined sample size adjustment options defined in fishery_iss()}
}
\value{
vector of resampled elements
}
\description{
Modified generic_sample() function to allow more flexibility based on user-defined options for adjusting haul-level sample size.
}
