% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iss.R
\name{iss}
\alias{iss}
\title{Calculate length or age composition input sampled size (iss)}
\usage{
iss(rss, freq_data, boot.length, boot.age)
}
\arguments{
\item{rss}{iterated length or age composition realized sample size}

\item{freq_data}{length or age frequency input dataframe}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.age}{Boolean. Resample ages w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-age}
}
\value{
iss
}
\description{
Function to calculate length or age composition iss (harmonic mean of realized sample sizes) and add nominal sample
size (the number of lengths or ages actually measured) and the number of sampled hauls to output.
}
