% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iss_length.R
\name{iss_length}
\alias{iss_length}
\title{Calculate length composition input sampled size (iss)}
\usage{
iss_length(rss_length, lfreq_data)
}
\arguments{
\item{rss_length}{iterated length composition realized sample size}

\item{lfreq_data}{length frequency input dataframe}
}
\description{
Function to calculate length composition iss (harmonic mean of realized sample sizes) and add nominal sample
size (the number of lengths actually measured) and the number of sampled hauls to output.
}
