% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_length.R
\name{plus_length}
\alias{plus_length}
\title{Plus group for length data}
\usage{
plus_length(length_DT, plus_len = NULL)
}
\arguments{
\item{length_DT}{length frequency input data frame.}

\item{plus_len}{If set at a value, computes length expansion with a plus-length group (default = FALSE)}
}
\value{
Data frame of length adapted with plus group
}
\description{
Creates a plus group
}
