% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reboot_length_N.R
\name{reboot_length_N}
\alias{reboot_length_N}
\title{Change the haul-level sample size of resampled lengths}
\usage{
reboot_length_N(length_DT, new_length_N = NULL)
}
\arguments{
\item{length_DT}{length frequency input data frame output from boot_length().}

\item{new_length_N}{user-defined sample size adjustment options defined in fishery_iss()}
}
\value{
length frequency data table including resampled length frequencies as LENGTH and associated data
}
\description{
Function similar to boot_length() to allow resampling of resampled length data based on user-defined options for adjusting haul-level sample size.
}
