% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss}
\alias{rss}
\title{Calculate length or age composition realized sampled size (rss)}
\usage{
rss(sim_props, og_props)
}
\arguments{
\item{sim_props}{list of replicated abundance at length or age}

\item{og_props}{original abundance at length or age}
}
\description{
Function to calculate rss for length or age composition following McAllister and Ianelli 1997.
Follows loosely comp_stats.R rss_length() from surveyISS package.
}
