% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_length.R
\name{rss_length}
\alias{rss_length}
\title{Calculate length composition realized sampled size (rss)}
\usage{
rss_length(sim_length_props, og_length_props)
}
\arguments{
\item{sim_length_props}{list of replicated abundance at length}

\item{og_length_props}{original abundance at length}
}
\description{
Function to calculate rss for length composition following McAllister and Ianelli 1997.
Follows loosely comp_stats.R rss_length() from surveyISS package.
}
