% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrate_iss.R
\name{administrate_iss}
\alias{administrate_iss}
\title{Administrate ISS}
\usage{
administrate_iss(freq_data)
}
\arguments{
\item{freq_data}{length or age frequency input data frame}
}
\value{
Data frame of input sample size by year
}
\description{
Administrate ISS
}
