% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_haul.R
\name{boot_haul}
\alias{boot_haul}
\title{Resample hauls w/replacement.}
\usage{
boot_haul(resampled_trips)
}
\arguments{
\item{resampled_trips}{resampled trips input data table. This is a truncated data table compared to length_DT. It has a column for YEAR, CRUISE, HAUL_JOIN only and one row per unique HAUL_JOIN.}
}
\value{
data table of resampled hauls by year
}
\description{
Resample hauls w/replacement.
}
