% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_length.R
\name{boot_length}
\alias{boot_length}
\title{Resample length frequency data w/replacement.}
\usage{
boot_length(length_DT)
}
\arguments{
\item{length_DT}{length frequency input data table. Can be the output object from boot_trip() if boot.trip == true, or boot_haul() if boot.haul == true, or both}
}
\value{
length frequency data table including resampled length frequencies as LENGTH and associated data
}
\description{
Resample length frequency data w/replacement.
}
