% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_trip.R
\name{boot_trip}
\alias{boot_trip}
\title{Resample trips w/replacement, changing the number of trips if desired}
\usage{
boot_trip(freq_data, new_trip_N = NULL)
}
\arguments{
\item{freq_data}{length or age frequency input data frame}

\item{new_trip_N}{user-defined sample size adjustment options defined in fishery_iss()}
}
\value{
frequency data table including resampled trips as TRIP_JOIN and associated data
}
\description{
Resample trips w/replacement, changing the number of trips if desired
}
