% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute fishery ISS statistics}
\usage{
compute_stats(sim_props = NULL, og_props = NULL, freq_data = NULL, iters = 1)
}
\arguments{
\item{sim_props}{list of replicated abundance at length or age}

\item{og_props}{original abundance at length or age (computed with data that has not been resampled)}

\item{freq_data}{length or age frequency input data frame}

\item{iters}{number of iterations}
}
\value{
list of dataframes for realized sample size by replicate (.rss for length or age composition),
input sample size by year (.iss),
bias in resampled comp data compared to original values (.bias for length or age composition)
}
\description{
Wrapper function to compute statistics of bootstrap resampling of length or age composition.
}
