% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute fishery ISS statistics}
\usage{
compute_stats(sim_length_props, og_length_props, lfreq_data)
}
\arguments{
\item{sim_length_props}{list of replicated abundance at length}

\item{og_length_props}{original abundance at length (computed with data that has not been resampled)}

\item{lfreq_data}{length frequency input dataframe}
}
\value{
list of dataframes for realized sample size by replicate (.rss_length for length composition),
input sample size by year? (.iss_length for length composition),
bias in resampled comp data compared to original values (.bias_length for length composition)
}
\description{
Wrapper function to compute statistics of bootstrap resampling of length composition.
}
