% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute fishery ISS statistics}
\usage{
compute_stats(
  sim_props = NULL,
  og_props = NULL,
  freq_data = NULL,
  iters = 1,
  boot.length = NULL,
  save_data_frame = FALSE
)
}
\arguments{
\item{sim_props}{list of replicated abundance at length or age}

\item{og_props}{original abundance at length or age (computed with data that has not been resampled)}

\item{freq_data}{length or age frequency input data frame}

\item{iters}{number of iterations}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{save_data_frame}{Boolean. Save freq_data data frame object that has been modified (e.g., filtered for minimum sample size) and used in ISS calculation? This may be useful for e.g. calculating haul or trip samples rates}
}
\value{
list of dataframes for realized sample size by replicate (.rss for length or age composition),
input sample size by year (.iss),
bias in resampled comp data compared to original values (.bias for length or age composition)
}
\description{
Wrapper function to compute statistics of bootstrap resampling of length or age composition.
}
