% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_props.R
\name{expand_props}
\alias{expand_props}
\title{Expand species-specific sample level proportions-at-length or -age to population level}
\usage{
expand_props(
  freq_data,
  species_code,
  area_code,
  expand.by.sampling.strata = FALSE,
  expand_using_weighting_factors = TRUE
)
}
\arguments{
\item{freq_data}{length or age frequency input data frame}

\item{species_code}{species specific number code to apply the expansion for. Conditional statements below activate species-specific expansions.}

\item{area_code}{area specific character code to apply the conditional expansion below.}

\item{expand.by.sampling.strata}{MAKE THIS GENERIC IN FUTURE. expand by observer sampling strata? If TRUE, then an additional weighting factor is calculated and applied to WEIGHT1 based on the number of fish caught in each sampling stratum.}

\item{expand_using_weighting_factors}{MAKE THIS GENERIC IN FUTURE. expand using weighting factors? If TRUE, then then "WEIGHT2" and "WEIGHT4" are applied.}
}
\description{
This is intended to be a large function that consists of many conditional statements, each of which pertains to a unique species/area expansion method. The internals of the conditional statements can eventually turn into many species/area/length/age specific functions housed in another script. Takes length or age frequency data and expands it into fished population proportions at length or age using the weightings from the relevant data. The relevant data can be the og data or bootstrapped data.
}
