% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_iss.R
\name{fishery_iss}
\alias{fishery_iss}
\title{Fishery input sample size function}
\usage{
fishery_iss(
  iters = 1,
  lfreq_data,
  yrs = NULL,
  boot.trip = FALSE,
  boot.haul = FALSE,
  boot.length = FALSE,
  expand.by.sampling.strata = FALSE,
  expand_using_weighting_factors = expand_using_weighting_factors
)
}
\arguments{
\item{iters}{number of iterations}

\item{lfreq_data}{length frequency input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{boot.trip}{Boolean. Resample trips w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.haul}{Boolean. Resample hauls w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{expand.by.sampling.strata}{expand by observer sampling strata? If TRUE, then an additional weighting factor is calculated and applied to WEIGHT1 based on the number of fish caught in each sampling stratum.}

\item{expand_using_weighting_factors}{expand using weighting factors? If TRUE, then then "WEIGHT2" and "WEIGHT4" are applied.}
}
\value{
Dataframe of input sample size by year
}
\description{
Bootstrap data sources to replicate EBS Pcod fishery observer measurements of length composition
for computation of input sample size.
Follows loosely srvy_iss.R from surveyISS package.
}
