% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_props.R
\name{fishery_props}
\alias{fishery_props}
\title{Resample data and compute proportions-at-length or -age population expansion}
\usage{
fishery_props(
  length_based,
  freq_data,
  yrs = NULL,
  boot.trip = FALSE,
  boot.haul = FALSE,
  boot.length = FALSE,
  boot.age = FALSE,
  expand.by.sampling.strata = FALSE,
  expand_using_weighting_factors = TRUE
)
}
\arguments{
\item{length_based}{Boolean. If TRUE, then calculate length iss. if FALSE, then calculate age iss.}

\item{freq_data}{length or age frequency input data frame}

\item{yrs}{any year filter >= (default = NULL)}

\item{boot.trip}{Boolean. Resample trips w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length or -age}

\item{boot.haul}{Boolean. Resample hauls w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length or -age}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.age}{Boolean. Resample ages w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-age}

\item{expand.by.sampling.strata}{expand by observer sampling strata? If TRUE, then an additional weighting factor is calculated and applied to WEIGHT1 based on the number of fish caught in each sampling stratum.}

\item{expand_using_weighting_factors}{expand using weighting factors? If TRUE, then then "WEIGHT2" and "WEIGHT4" are applied.}
}
\value{
List of a dataframe of annual population proportions-at-length .lpop.
}
\description{
Follows loosely srvy_comps.R from surveyISS package.
}
