% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratify.R
\name{post_stratify}
\alias{post_stratify}
\title{Post-stratify ISS}
\usage{
post_stratify(lfreq_data, post_strata)
}
\arguments{
\item{lfreq_data}{length frequency input dataframe}

\item{post_strata}{list(strata = character string, nested = Boolean) if NULL, then no post stratification. The first element of post_strata is a character string with name(s) of post strata type. Accepted types are "GEAR", etc. These must be column name(s) in the data frame where each row has an entry and there are no NAs. The second element in post_strata allows for nested post-strata (e.g., "GEAR", "SEX") that are in order of the desired nested hierarchy, then the final results will be a nested list in order of the listed post-strata as well. Perhaps there should be an error message for attempts with strata names that are not in lfreq_data or it does exist, but there are NAs in it.}
}
\value{
list of input sample size by post-strata and year
}
\description{
Apply user-requested post-stratification for one or up to two nested post-stratum
}
