% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratify.R
\name{post_stratify}
\alias{post_stratify}
\title{Stratify ISS results by post-stratum.}
\usage{
post_stratify(lfreq_data)
}
\arguments{
\item{lfreq_data}{length frequency input dataframe, not necessarily a data.table or tidytable object yet, hence a different name than subsequent functions (length_DT)}
}
\value{
Dataframe of input sample size by year
}
\description{
Stratify ISS results by post-stratum.
}
