% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratify.R
\name{post_stratify}
\alias{post_stratify}
\title{Post-stratify ISS}
\usage{
post_stratify(
  species_code,
  area_code,
  length_based = TRUE,
  iters = 1,
  freq_data,
  bin = NULL,
  plus_len = NULL,
  post_strata = NULL,
  minimum_sample_size = NULL,
  new_trip_N = NULL,
  new_length_N = NULL,
  max_length,
  boot.trip = FALSE,
  boot.haul = FALSE,
  boot.length = TRUE,
  boot.age = FALSE,
  expand.by.sampling.strata = FALSE,
  expansion_factors = NULL
)
}
\arguments{
\item{species_code}{species number code. Used for specific expansion. This is not (yet) an input data filter, it is only for output naming convention and to condition on expansion method.}

\item{area_code}{area character code. Used for specific expansion. This is not (yet) an input data filter, it is only for output naming convention and to condition on expansion method.}

\item{length_based}{Boolean. If TRUE, then calculate length iss. if FALSE, then calculate age iss.}

\item{iters}{number of iterations}

\item{freq_data}{length or age frequency input data frame}

\item{bin}{bin size (NULL = 1 cm), also can use custom length bins through defining vector of upper length bin limits, i.e., c(5, 10, 20, 35, 60), plus length group will automatically be populated and denoted with the largest defined bin + 1 (i.e., 61 for provided example)}

\item{plus_len}{If set at a value, computes length expansion with a plus-length group (default = FALSE)}

\item{post_strata}{list(strata = character string, nested = Boolean) if NULL, then no post stratification. The first element of post_strata is a character string with name(s) of post strata type. Accepted types are "GEAR", etc. These must be column name(s) in the data frame where each row has an entry and there are no NAs. The second element in post_strata allows for nested post-strata (e.g., "GEAR", "SEX") that are in order of the desired nested hierarchy, then the final results will be a nested list in order of the listed post-strata as well. Perhaps there should be an error message for attempts with strata names that are not in freq_data or it does exist, but there are NAs in it.}

\item{minimum_sample_size}{list(resolution = character string, size = integer). If NULL, then no minimum sample size. If not NULL, The sample size at the chosen resolution (must be column in freq_data, e.g., YAGM_SFREQ for EBS Pcod) for which to filter out data that does not meet the minimum sample size requested. Example: minimum_sample_size = list(resolution = "YAGM_SFREQ", size = 30). Note that this filters to keep only samples GREATER than 30 at the YAGM resolution.}

\item{new_trip_N}{list(type = character ("value", or "proportion"), amount = numeric). If NULL, then the number of trips resampled equals the number actually sampled. If not NULL, then the number of trips is changed by type (acceptable entries are "value", or "proportion") at an amount (acceptable entries are a proportion). E.g., new_trip_N = list(type = "proportion", amount = .50). Only proportion is coded so far as this makes the most sense.}

\item{new_length_N}{list(type = character ("value", or "proportion"), bound = NULL or character ("minimum", or "maximum") amount = numeric). If NULL, then the number of length samples resampled in a haul equals the number actually sampled. If not NULL, then the haul-level number of samples is changed by type (acceptable entries are "value", or "proportion") at an amount (acceptable entries are an integer number or proportion). E.g., new_length_N = list(type = "value", bound = NULL, amount = 20). The amount can be less than or greater than the true N. Note that if it is either less than or greater than, the minimum of the two is chosen for N (this is still under consideration and may change in the future). The true N is taken to be... DOES THIS ONLY MATTER IF WE NEED A THRESHOLD TO DECIDE WHEN TO TAKE THE MAX?? The samples are still drawn with replacement. Warning: cannot do this if post_strata = "SEX"}

\item{max_length}{integer or "data_derived" (default). Provides the maximum length bin to record proportions-at-length. Integer values must be equal to or greater than the maximum length observed in the data (this will be changed once a plus-group option is coded in the future). "data_derived" forces the maximum length to be calculated from the data.}

\item{boot.trip}{Boolean. Resample trips w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length or -age}

\item{boot.haul}{Boolean. Resample hauls w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length or -age}

\item{boot.length}{Boolean. Resample lengths w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-length}

\item{boot.age}{Boolean. Resample ages w/replacement? (default = FALSE). FALSE to all three boots will return og proportions-at-age}

\item{expand.by.sampling.strata}{expand by observer sampling strata? If TRUE, then an additional weighting factor is calculated and applied to WEIGHT1 based on the number of fish caught in each sampling stratum.}

\item{expansion_factors}{expansion weighting factors to apply to the proportions. If NULL, then no expansion factors are applied. Otherwise, the conditional options coded in expand_props.R are "haul_numbers" or "haul_numbers" and "month_numbers". Consider improving/generalizing this by calling it expansion_weighting_factors = list(type = c("weight", "number"), factors = c("haul", "area", "month", "gear", etc.) to give the user the option of what aspects (columns) of the data to expand by and do it by weight of fish or number of fish in those categories.}
}
\value{
list of input sample size by post-strata and year
}
\description{
Apply user-requested post-stratification for one or up to two nested post-stratum
}
